+++ spec/db_fuel/library/dbee/range_spec.rb
@@ -62,6 +62,12 @@ describe DbFuel::Library::Dbee::Range do
       subject.perform(output, payload)
     end
 
+    specify 'output contains number of records' do
+      string_summary = output.outs.first
+
+      expect(string_summary.read).to include("Loading 2 record(s) into #{register}")
+    end
+
     specify 'payload register has data' do
       records = payload[register]